Adding cakephp 1.3.4

+++ CakePlate/cake/libs/view/elements/sql_dump.ctp
@@ -20,9 +20,10 @@
 if (!class_exists('ConnectionManager') || Configure::read('debug') < 2) {
 	return false;
 }
+$noLogs = !isset($logs);
+if ($noLogs):
+	$sources = ConnectionManager::sourceList();
 
-$sources = ConnectionManager::sourceList();
-if (!isset($logs)):
 	$logs = array();
 	foreach ($sources as $source):
 		$db =& ConnectionManager::getDataSource($source);
@@ -33,22 +34,28 @@ if (!isset($logs)):
 	endforeach;
 endif;
 
-foreach ($logs as $source => $logInfo):
-	$text = $logInfo['count'] > 1 ? 'queries' : 'query';
-	printf(
-		'<table class="cake-sql-log" id="cakeSqlLog_%s" summary="Cake SQL Log" cellspacing="0" border = "0">',
-		preg_replace('/[^A-Za-z0-9_]/', '_', uniqid(time(), true))
-	);
-	printf('<caption>(%s) %s %s took %s ms</caption>', $source, $logInfo['count'], $text, $logInfo['time']);
-?>
-<thead>
-	<tr><th>Nr</th><th>Query</th><th>Error</th><th>Affected</th><th>Num. rows</th><th>Took (ms)</th></tr>
-</thead>
-<tbody>
-<?php
-	foreach ($logInfo['log'] as $k => $i) :
-		echo "<tr><td>" . ($k + 1) . "</td><td>" . h($i['query']) . "</td><td>{$i['error']}</td><td style = \"text-align: right\">{$i['affected']}</td><td style = \"text-align: right\">{$i['numRows']}</td><td style = \"text-align: right\">{$i['took']}</td></tr>\n";
+if ($noLogs || isset($_forced_from_dbo_)):
+	foreach ($logs as $source => $logInfo):
+		$text = $logInfo['count'] > 1 ? 'queries' : 'query';
+		printf(
+			'<table class="cake-sql-log" id="cakeSqlLog_%s" summary="Cake SQL Log" cellspacing="0" border = "0">',
+			preg_replace('/[^A-Za-z0-9_]/', '_', uniqid(time(), true))
+		);
+		printf('<caption>(%s) %s %s took %s ms</caption>', $source, $logInfo['count'], $text, $logInfo['time']);
+	?>
+	<thead>
+		<tr><th>Nr</th><th>Query</th><th>Error</th><th>Affected</th><th>Num. rows</th><th>Took (ms)</th></tr>
+	</thead>
+	<tbody>
+	<?php
+		foreach ($logInfo['log'] as $k => $i) :
+			echo "<tr><td>" . ($k + 1) . "</td><td>" . h($i['query']) . "</td><td>{$i['error']}</td><td style = \"text-align: right\">{$i['affected']}</td><td style = \"text-align: right\">{$i['numRows']}</td><td style = \"text-align: right\">{$i['took']}</td></tr>\n";
+		endforeach;
+	?>
+	</tbody></table>
+	<?php 
 	endforeach;
+else:
+	echo '<p>Encountered unexpected $logs cannot generate SQL log</p>';
+endif;
 ?>
-</tbody></table>
-<?php endforeach; ?>
